sleepapp/requirements.txt
Django==5.0.1
django-rest-framework==0.1.0
psycopg2-binary
django-cors-headers
numpy